refactor: extract validation schemas to shared layer
Eliminate the last lib/ → features/ dependency violation by extracting
Zod schemas from features/ into src/shared/schemas/. This allows
src/lib/api/openapi.ts to import createLearningPlanSchema from shared
instead of features, fixing the architecture layer violation.

Changes:
- Create src/shared/schemas/plan-validation.schemas.ts with SKILL_LEVEL_ENUM, LEARNING_STYLE_ENUM, weeklyHoursSchema
- Create src/shared/schemas/pdf-validation.schemas.ts with pdfExtractedSectionSchema, pdfPreviewEditSchema
- Create src/shared/schemas/learning-plans.schemas.ts with createLearningPlanSchema, createLearningPlanObject, topicSchema
- Update src/features/plans/validation/shared.ts to re-export from shared
- Update src/features/pdf/validation/pdf.ts to re-export from shared
- Update src/features/plans/validation/learningPlans.ts to re-export from shared
- Update src/lib/api/openapi.ts to import from shared instead of features
- Remove openapi.ts ESLint exception from eslint.config.mjs
