+++ src/features/pdf/validation/pdf.ts
@@ -7,13 +7,12 @@ export type {
   TruncationData,
 } from './pdf.types';
 
-import {
-  LEARNING_STYLE_ENUM,
-  NOTES_MAX_LENGTH,
-  SKILL_LEVEL_ENUM,
-  TOPIC_MAX_LENGTH,
-  weeklyHoursSchema,
-} from '@/features/plans/validation/shared';
+import { TOPIC_MAX_LENGTH } from '@/shared/constants/learning-plans';
+export {
+  pdfExtractedSectionSchema,
+  pdfPreviewEditSchema,
+} from '@/shared/schemas/pdf-validation.schemas';
+import { pdfExtractedSectionSchema } from '@/shared/schemas/pdf-validation.schemas';
 
 export const pdfExtractionRequestSchema = z
   .object({
@@ -23,31 +22,11 @@ export const pdfExtractionRequestSchema = z
   })
   .strict();
 
-export const pdfExtractedSectionSchema = z
-  .object({
-    title: z.string().trim().min(1).max(200),
-    content: z.string().trim().max(5000),
-    level: z.number().int().min(1).max(5),
-    suggestedTopic: z.string().trim().max(200).optional(),
-  })
-  .strict();
-
 export const pdfExtractedContentSchema = z
   .object({
     mainTopic: z.string().trim().min(3).max(TOPIC_MAX_LENGTH),
     sections: z.array(pdfExtractedSectionSchema).min(1).max(50),
     confidence: z.enum(['high', 'medium', 'low']),
-  })
-  .strict();
-
-export const pdfPreviewEditSchema = z
-  .object({
-    mainTopic: z.string().trim().min(3).max(TOPIC_MAX_LENGTH),
-    sections: z.array(pdfExtractedSectionSchema).min(1).max(50),
-    skillLevel: SKILL_LEVEL_ENUM.optional(),
-    weeklyHours: weeklyHoursSchema.optional(),
-    learningStyle: LEARNING_STYLE_ENUM.optional(),
-    notes: z.string().trim().max(NOTES_MAX_LENGTH).optional(),
   })
   .strict();
